Support full object editing. It is now possible to write any PDF snippet that can be parsed by PdfTokenizer in the value field of an object, and PoDoFoBrowser should do the right thing. This includes container objects like arrays and dictionaries, and also includes references. Changing types works fine - eg replacing a value `4' with a value `<< /Type /Pages /Name (fred) /Arr [ 1 2 3 4 ] >>' will populate the newly created subtree. The reverse is also true.
Note that since the value of an array or dictionary is shown truncated, editing them in place will lose data. This will be fixed as things progress. It's entirely safe to use the browser to change elements, of course.

A known stream display bug exists in this and earlier revisions. If a stream's dictionary has been shown to the user before the stream is delay-loaded and the stream has an indirect length key, podofo replaces this with a direct length. This change breaks the node references and will cause a crash. A fix will be checked in shortly.

git-svn-id: svn://svn.code.sf.net/p/podofo/code@380 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5