Fixed PdfStreamedDocument. Correct PDF files are now generated and PdfStreamedDocument does now apply compression filters on streams.
LargeTest is now about twice as fast when using PdfStreamedDocument as when using PdfDocument.

I will add support for Flate Predictors and try to solve the still existing problem that some destructors are quite likely to 
throw exceptions.

But as soon as this is done I would vote for a release. Comments? What's your plans?




git-svn-id: svn://svn.code.sf.net/p/podofo/code@447 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5