More progess towards cleaning up direct/indirect distinction, separating usage of PdfVariant and PdfObject. It's slow going as everything is incredibly interdependent and makes a lot of assumptions about what it's dealing with.
This code does not build.

Additionally, there appear to be problems where a class needs to resolve an
indirect reference, but has no information about the document it's associated
with. Perhaps indirect references need a handle to the document / vecobjects?
Or maybe at the PdfVariant level? Issues all around there that need looking into.


git-svn-id: svn://svn.code.sf.net/p/podofo/code@487 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5