FIXED: Mantis issue 0000004 I fixed the two reported problems in SVN trunk. PdfEncodingObjectFactory creates now a PdfIdentityEncoding when it finds an /Identity-H key. Also, we support DescendantFonts that are in an indirect object like in your example created by MS Office.
Still, I think you will not be able to create a PdfFontCID from an existing object. The class does not yet support this and several fixes will be necessary to support this. Please, create another issue for this, maybe with a description what you need exactly or what you want to do.



git-svn-id: svn://svn.code.sf.net/p/podofo/code@1392 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5