Patch by Jonathan Verner to optimize laoding of PDFs with object streams:
when trying to use PoDoFo to parse large files (e.g. the PDF spec),
PdfMemDocument took forever to load. When I tried running, e.g.
podofocountpages on the PDF spec after approx. ten minutes it wasn't
finished yet.

After looking at it, I found out that most of the time
is spent in

  PdfObjectStreamparserObject::ReadObjectsFromStream

in the while cycle.

First, on line 98 (PdfObjectStreamParserObject.cpp)
GetObject is called. If the m_vecObjects vector is not sorted, the
method first calls sort to sort it.
Then, on line 102, a new object is pushed_back into the vector,
which invalidates the sorting so when GetObject is called again,
it has to sort the list again --- Quite time consuming.

I have tried replacing the push_back on line 102 by a new
method insert_sorted, which inserts the element into
the vector so that it stays sorted (if it was sorted in the first place)
and this has led to a dramatic improvement: now podofocountpages
finishes in a reasonable amount of time:

$ time podofocountpages PDF32000_2008.pdf
PDF32000_2008.pdf:      756

real    0m20.235s
user    0m16.809s
sys     0m0.260s


Attached is my proposed patch. What do you think about it?



git-svn-id: svn://svn.code.sf.net/p/podofo/code@1457 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5