PdfPainter/MM: Change the way of drawing primitives
This removes Draw and Fill functions of primitives, and
replaces them to two-way drawing:
a) path construction
b) draw the path with Stroke()/Fill()/StrokeAndFill()/EndPath()/Clip(), respectively

Current path is also remembered and can be read by PdfPainter::GetCurrentPath(),
same as returned back with PdfPainter::AddRawCommands().

As an example, previous
   DrawRect()
can be replaced with
   Rectangle(); Stroke();
sequence, and
   FillRect()
with
   Rectangle(); Fill();
sequence.

Advantages:
- one can fill and stroke in one call (which PDF supports) with
     Rectangle(); FillAndStroke();
  sequence
- defined path can be used for clipping as well


git-svn-id: svn://svn.code.sf.net/p/podofo/code@1541 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5