ADDED: Patch by Matthew Brincke: Documentation cleanup for "ErrOk on success" return value where none is given
Clean up documentation for methods with return type "void" and a constructor
where a return of "ErrOk" is still mentioned: remove such mentions, also
clarify some parameter and method documentation and correct the constructor's
documentation. Also remove of @returns in favor of \returns and correction of
return-value documentation (both in PdfVariant.h).


git-svn-id: svn://svn.code.sf.net/p/podofo/code@1677 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5